@@ -6,7 +6,8 @@
     "@babel/core": "^7.16.0",
     "@fluentui/react-charting": "^5.24.37",
     "@fluentui/react-charts": "0.0.0-nightly-20251010-0407.1",
-    "@fluentui/react-components": "^9.21.0",
+    "@fluentui/react-components": "^9.72.2",
+    "@fluentui/react-menu": "^9.20.1",
     "@fluentui/react-portal-compat": "9.0.176",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
     "@svgr/webpack": "^5.5.0",
@@ -45,10 +46,10 @@
     "postcss-normalize": "^10.0.1",
     "postcss-preset-env": "^7.0.1",
     "prompts": "^2.4.2",
-    "react": "^17.0.2",
+    "react": "^19.0.0",
     "react-app-polyfill": "^3.0.0",
     "react-dev-utils": "^12.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.0.0",
     "react-plotly.js": "^2.6.0",
     "react-refresh": "^0.11.0",
     "resolve": "^1.20.0",
@@ -93,15 +94,15 @@
   "devDependencies": {
     "@playwright/test": "^1.49.1",
     "@types/node": "^22.10.7",
-    "@types/react": "17.0.44",
-    "@types/react-dom": "17.0.15",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
     "parse-svg-path": "^0.1.2"
   },
   "resolutions": {
-    "@types/react": "17.0.44",
-    "@types/react-dom": "17.0.15",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@types/react": "19.0.0",
+    "@types/react-dom": "19.0.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "@fluentui/chart-utilities": "^1.1.33"
   },
   "jest": {
